+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- fix layer order in set active layer tool
+
 ## [3.9.4] - 2023-02-08
 
 - fix plugin packaging